Move fetch ops over to op_crates/fetch
