+++ op_crates/fetch/Cargo.toml
@@ -15,3 +15,5 @@ path = "lib.rs"
 
 [dependencies]
 deno_core = { version = "0.57.0", path = "../../core" }
+reqwest = { version = "0.10.8", default-features = false, features = ["rustls-tls", "stream", "gzip", "brotli"] }
+serde = { version = "1.0.116", features = ["derive"] }
